+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/ ).
 and this project adheres to [Semantic Versioning](http://semver.org/ ).
 
+## Unreleased
+
+### Changed
+- Added installed module migration paths in `getMigrationPaths` method
+
 ## [2.3.2] - 2020-03-19
 
 ### Added